Convert remaining Cli commands to IFalloutCommand and collapse Program (#392)
Convert complete, get-configuration, add-package, update, secrets, cake-convert, cake-clean,
setup, and the navigation cluster into real IFalloutCommand types (async ExecuteAsync, DI-injected
services), register them directly, and delete the DelegateCommand shims + every Program.X.cs
partial. Program is now a thin entry point (DI wiring + root resolution). Commands and new CLI
interfaces are internal, matching the merged foundation. Derived from main's current handlers, so
the .slnx setup support, net10.0 target, and nuke->fallout rebrand are preserved.
