+++ src/Fallout.Cli/Commands/CompleteCommand.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading.Tasks;
 using System.Collections.Generic;
 using System.Linq;
 using Fallout.Common;
@@ -7,13 +8,21 @@ using Fallout.Common.Utilities;
 using Fallout.Utilities.Text.Yaml;
 using static Fallout.Common.Constants;
 
-namespace Fallout.Cli;
+namespace Fallout.Cli.Commands;
 
-partial class Program
+/// <summary>
+/// <c>fallout :complete</c>: emits shell-completion candidates for the partially typed command line.
+/// </summary>
+internal sealed class CompleteCommand : IFalloutCommand
 {
     private const string CommandName = "fallout";
 
-    public static int Complete(string[] args, AbsolutePath rootDirectory, AbsolutePath buildScript)
+    public string Name => "complete";
+
+    public Task<int> ExecuteAsync(string[] args, AbsolutePath rootDirectory, AbsolutePath buildScript)
+        => Task.FromResult(Execute(args, rootDirectory, buildScript));
+
+    private int Execute(string[] args, AbsolutePath rootDirectory, AbsolutePath buildScript)
     {
         if (rootDirectory == null)
             return 0;
