--- src/Fallout.Cli/Program.AddPackage.cs
@@ -1,63 +1,0 @@
-﻿using System;
-using System.Linq;
-using Fallout.Common;
-using Fallout.Common.Execution;
-using Fallout.Common.IO;
-using Fallout.Solutions;
-using Fallout.Common.Tooling;
-using Fallout.Common.Tools.DotNet;
-
-namespace Fallout.Cli;
-
-partial class Program
-{
-    public const string PACKAGE_TYPE_DOWNLOAD = "PackageDownload";
-    public const string PACKAGE_TYPE_REFERENCE = "PackageReference";
-
-    public static int AddPackage(string[] args, AbsolutePath rootDirectory, AbsolutePath buildScript)
-    {
-        PrintInfo();
-        Logging.Configure();
-        Telemetry.AddPackage();
-        ProjectModelTasks.Initialize();
-
-        var packageId = args.ElementAt(0);
-        var packageVersion =
-            (EnvironmentInfo.GetNamedArgument<string>("version") ??
-             args.ElementAtOrDefault(1) ??
-             NuGetVersionResolver.GetLatestVersion(packageId, includePrereleases: false).GetAwaiter().GetResult() ??
-             NuGetPackageResolver.GetGlobalInstalledPackage(packageId, version: null, packagesConfigFile: null)?.Version.ToString())
-            .NotNull("packageVersion != null");
-
-        var configuration = GetConfiguration(buildScript, evaluate: true);
-        var buildProjectFile = configuration[BUILD_PROJECT_FILE];
-        Host.Information($"Installing {packageId}/{packageVersion} to {buildProjectFile} ...");
-        AddOrReplacePackage(packageId, packageVersion, PACKAGE_TYPE_DOWNLOAD, buildProjectFile);
-        DotNetTasks.DotNet($"restore {buildProjectFile}");
-
-        var installedPackage = NuGetPackageResolver.GetGlobalInstalledPackage(packageId, packageVersion, packagesConfigFile: null)
-            .NotNull("installedPackage != null");
-        var hasToolsDirectory = installedPackage.Directory.GlobDirectories("tools").Any();
-        if (!hasToolsDirectory)
-            AddOrReplacePackage(packageId, packageVersion, PACKAGE_TYPE_REFERENCE, buildProjectFile);
-
-        Host.Information($"Done installing {packageId}/{packageVersion} to {buildProjectFile}");
-        return 0;
-    }
-
-    internal static void AddOrReplacePackage(string packageId, string packageVersion, string packageType, string buildProjectFile)
-    {
-        var buildProject = ProjectModelTasks.ParseProject(buildProjectFile).NotNull();
-
-        var previousPackage = buildProject.Items.SingleOrDefault(x => x.EvaluatedInclude == packageId);
-        if (previousPackage != null)
-            buildProject.RemoveItem(previousPackage);
-
-        var packageDownloadItem = buildProject.AddItem(packageType, packageId).Single();
-        packageDownloadItem.Xml.AddMetadata(
-            "Version",
-            packageType == PACKAGE_TYPE_REFERENCE ? packageVersion : $"[{packageVersion}]",
-            expressAsAttribute: true);
-        buildProject.Save();
-    }
-}
