--- src/Fallout.Cli/Program.Setup.cs
@@ -1,281 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Xml;
-using System.Xml.Linq;
-using Fallout.Common;
-using Fallout.Common.Execution;
-using Fallout.Common.IO;
-using Fallout.Common.Tooling;
-using Fallout.Common.Utilities;
-using Fallout.Common.Utilities.Collections;
-using Spectre.Console;
-using static Fallout.Common.Constants;
-using static Fallout.Common.EnvironmentInfo;
-using static Fallout.Common.Tooling.ProcessTasks;
-using static Fallout.Common.Utilities.TemplateUtility;
-
-namespace Fallout.Cli;
-
-partial class Program
-{
-    // ReSharper disable InconsistentNaming
-
-    private const string TARGET_FRAMEWORK = "net10.0";
-    private const string PROJECT_KIND = "9A19103F-16F7-4668-BE54-9A1E7A4F7556";
-
-    // ReSharper disable once CognitiveComplexity
-    public static int Setup(string[] args, AbsolutePath rootDirectory, AbsolutePath buildScript)
-    {
-        PrintInfo();
-        Logging.Configure();
-        Telemetry.SetupBuild();
-
-        AnsiConsole.WriteLine();
-        AnsiConsole.MarkupLine("[bold]Let's setup a new build![/]");
-        AnsiConsole.WriteLine();
-
-        #region Basic
-
-        var falloutLatestReleaseVersion = NuGetVersionResolver.GetLatestVersion(FalloutCommonPackageId, includePrereleases: false);
-        var falloutLatestPrereleaseVersion = NuGetVersionResolver.GetLatestVersion(FalloutCommonPackageId, includePrereleases: true);
-        var falloutLatestLocalVersion = NuGetPackageResolver.GetGlobalInstalledPackage(FalloutCommonPackageId, version: null, packagesConfigFile: null)
-            ?.Version.ToString();
-
-        if (rootDirectory == null)
-            rootDirectory = WorkingDirectory.FindParentOrSelf(x => x.ContainsDirectory(".git") || x.ContainsDirectory(".svn"));
-
-        if (rootDirectory == null)
-        {
-            Host.Warning("Could not find root directory. Falling back to working directory ...");
-            rootDirectory = WorkingDirectory;
-        }
-        ShowInput("deciduous_tree", "Root directory", rootDirectory);
-
-        var buildProjectName = PromptForInput("How should the project be named?", "_build");
-        ClearPreviousLine();
-        ShowInput("bookmark", "Build project name", buildProjectName);
-
-        var buildProjectRelativeDirectory = PromptForInput("Where should the project be located?", "./build");
-        ClearPreviousLine();
-        ShowInput("round_pushpin", "Build project location", buildProjectRelativeDirectory);
-
-        var falloutVersion = PromptForChoice("Which Fallout.Common version should be used?",
-            new[]
-                {
-                    ("latest release", falloutLatestReleaseVersion.GetAwaiter().GetResult()),
-                    ("latest prerelease", falloutLatestPrereleaseVersion.GetAwaiter().GetResult()),
-                    ("latest local", falloutLatestLocalVersion),
-                    ("same as global tool", typeof(Program).GetTypeInfo().Assembly.GetVersionText())
-                }
-                .Where(x => x.Item2 != null)
-                .Distinct(x => x.Item2)
-                .Select(x => (x.Item2, $"{x.Item2} ({x.Item1})")).ToArray());
-        ShowInput("gem_stone", "Fallout.Common version", falloutVersion);
-
-        var solutionFile = (AbsolutePath) PromptForChoice(
-            "Which solution should be the default?",
-            choices: new DirectoryInfo(rootDirectory)
-                .EnumerateFiles("*", SearchOption.AllDirectories)
-                .Where(x => x.FullName.EndsWithOrdinalIgnoreCase(".sln") || x.FullName.EndsWithOrdinalIgnoreCase(".slnx"))
-                .OrderByDescending(x => x.FullName)
-                .Select(x => (x, rootDirectory.GetRelativePathTo(x.FullName).ToString()))
-                .Concat((null, "None")).ToArray())?.FullName;
-        ShowInput("toolbox", "Default solution", solutionFile != null ? rootDirectory.GetRelativePathTo(solutionFile) : "<none>");
-
-        #endregion
-
-        #region Generation
-
-        var buildDirectory = rootDirectory / buildProjectRelativeDirectory;
-        var buildProjectFile = rootDirectory / buildProjectRelativeDirectory / buildProjectName + ".csproj";
-        var buildProjectGuid = Guid.NewGuid().ToString().ToUpper();
-
-        (rootDirectory / FalloutDirectoryName).CreateDirectory();
-
-        WriteBuildScripts(
-            scriptDirectory: WorkingDirectory,
-            rootDirectory);
-
-        WriteConfigurationFile(rootDirectory, solutionFile);
-
-        if (solutionFile != null)
-        {
-            var buildProjectFileRelative = solutionFile.Parent.GetWinRelativePathTo(buildProjectFile);
-            if (solutionFile.Extension.EqualsOrdinalIgnoreCase(".slnx"))
-            {
-                var solutionDocument = XDocument.Load(solutionFile);
-                UpdateSolutionXmlFileContent(solutionDocument, buildProjectFileRelative);
-
-                var settings = new XmlWriterSettings { OmitXmlDeclaration = true, Indent = true };
-                using var writer = XmlWriter.Create(solutionFile, settings);
-                solutionDocument.Save(writer);
-            }
-            else
-            {
-                var solutionFileContent = solutionFile.ReadAllLines().ToList();
-                UpdateSolutionFileContent(solutionFileContent, buildProjectFileRelative, buildProjectGuid, buildProjectName);
-                solutionFile.WriteAllLines(solutionFileContent, Encoding.UTF8);
-            }
-        }
-
-        buildProjectFile.WriteAllLines(
-            FillTemplate(
-                GetTemplate("_build.csproj"),
-                GetDictionary(
-                    new
-                    {
-                        RootDirectory = buildDirectory.GetWinRelativePathTo(rootDirectory),
-                        ScriptDirectory = buildDirectory.GetWinRelativePathTo(WorkingDirectory),
-                        TargetFramework = TARGET_FRAMEWORK,
-                        TelemetryVersion = Telemetry.CurrentVersion,
-                        FalloutVersion = falloutVersion,
-                    })));
-
-        (buildDirectory / "Directory.Build.props").WriteAllLines(GetTemplate("Directory.Build.props"));
-        (buildDirectory / "Directory.Build.targets").WriteAllLines(GetTemplate("Directory.Build.targets"));
-        (buildDirectory / "Build.cs").WriteAllLines(FillTemplate(GetTemplate("Build.cs")));
-        (buildDirectory / "Configuration.cs").WriteAllLines(GetTemplate("Configuration.cs"));
-
-        #endregion
-
-        ShowCompletion("Setup");
-
-        return 0;
-    }
-
-    internal static void UpdateSolutionFileContent(
-        List<string> content,
-        string buildProjectFileRelative,
-        string buildProjectGuid,
-        string buildProjectName)
-    {
-        if (content.Any(x => x.Contains(buildProjectFileRelative)))
-            return;
-
-        var globalIndex = content.IndexOf("Global");
-        Assert.True(globalIndex != -1, "Could not find a 'Global' section in solution file");
-
-        var projectConfigurationIndex = content.FindIndex(x => x.Contains("GlobalSection(ProjectConfigurationPlatforms)"));
-        if (projectConfigurationIndex == -1)
-        {
-            var solutionConfigurationIndex = content.FindIndex(x => x.Contains("GlobalSection(SolutionConfigurationPlatforms)"));
-            if (solutionConfigurationIndex == -1)
-            {
-                content.Insert(globalIndex + 1, "\tGlobalSection(SolutionConfigurationPlatforms) = preSolution");
-                content.Insert(globalIndex + 2, "\t\tDebug|Any CPU = Debug|Any CPU");
-                content.Insert(globalIndex + 3, "\t\tRelease|Any CPU = Release|Any CPU");
-                content.Insert(globalIndex + 4, "\tEndGlobalSection");
-
-                solutionConfigurationIndex = globalIndex + 1;
-            }
-
-            var endGlobalSectionIndex = content.FindIndex(solutionConfigurationIndex, x => x.Contains("EndGlobalSection"));
-
-            content.Insert(endGlobalSectionIndex + 1, "\tGlobalSection(ProjectConfigurationPlatforms) = postSolution");
-            content.Insert(endGlobalSectionIndex + 2, "\tEndGlobalSection");
-
-            projectConfigurationIndex = endGlobalSectionIndex + 1;
-        }
-
-        content.Insert(projectConfigurationIndex + 1, $"\t\t{{{buildProjectGuid}}}.Debug|Any CPU.ActiveCfg = Debug|Any CPU");
-        content.Insert(projectConfigurationIndex + 2, $"\t\t{{{buildProjectGuid}}}.Release|Any CPU.ActiveCfg = Release|Any CPU");
-
-        content.Insert(globalIndex,
-            $"Project(\"{{{PROJECT_KIND}}}\") = \"{buildProjectName}\", \"{buildProjectFileRelative}\", \"{{{buildProjectGuid}}}\"");
-        content.Insert(globalIndex + 1,
-            "EndProject");
-    }
-
-    internal static void UpdateSolutionXmlFileContent(XDocument content, string buildProjectFileRelative)
-    {
-        var solutionElement = content.Root;
-        Assert.True(solutionElement?.Name == "Solution", "Could not find a root 'Solution' element in solution file");
-
-        // file uses forward slashes for paths on every platform
-        var path = buildProjectFileRelative.Replace(oldChar: '\\', newChar: '/');
-
-        if (solutionElement.Elements("Project").Any(x => x.GetAttributeValue("Path").EqualsOrdinalIgnoreCase(path)))
-        {
-            return;
-        }
-
-        var projectElement = new XElement("Project", new XAttribute("Path", path));
-        projectElement.Add(new XElement("Build", new XAttribute("Project", value: false)));
-        solutionElement.Add(projectElement);
-    }
-
-    internal static string[] GetTemplate(string templateName)
-    {
-        return ResourceUtility.GetResourceAllLines<Program>($"templates.{templateName}");
-    }
-
-    internal static void WriteBuildScripts(
-        AbsolutePath scriptDirectory,
-        AbsolutePath rootDirectory)
-    {
-        (scriptDirectory / "build.sh").WriteAllLines(
-            FillTemplate(
-                GetTemplate("build.sh"),
-                tokens: GetDictionary(
-                    new
-                    {
-                        RootDirectory = scriptDirectory.GetUnixRelativePathTo(rootDirectory),
-                    })),
-            platformFamily: PlatformFamily.Linux);
-
-        (scriptDirectory / "build.ps1").WriteAllLines(
-            FillTemplate(
-                GetTemplate("build.ps1"),
-                tokens: GetDictionary(
-                    new
-                    {
-                        RootDirectory = scriptDirectory.GetWinRelativePathTo(rootDirectory),
-                    })),
-            platformFamily: PlatformFamily.Windows);
-
-        // .config/dotnet-tools.json pins Fallout.GlobalTools as a local tool so the thin shims
-        // (build.sh / build.ps1) can `dotnet tool restore` and `dotnet fallout` deterministically.
-        // Skip if the consumer already has a manifest — they may have other tools pinned and we
-        // don't want to clobber. They can add the `fallout.globaltools` entry manually.
-        var toolManifest = rootDirectory / ".config" / "dotnet-tools.json";
-        if (!toolManifest.FileExists())
-        {
-            (rootDirectory / ".config").CreateDirectory();
-            toolManifest.WriteAllLines(
-                FillTemplate(
-                    GetTemplate("dotnet-tools.json"),
-                    tokens: GetDictionary(
-                        new
-                        {
-                            FalloutCliVersion = typeof(Program).GetTypeInfo().Assembly.GetVersionText(),
-                        })));
-        }
-
-        MakeExecutable(scriptDirectory / "build.sh");
-
-        void MakeExecutable(AbsolutePath scriptFile)
-        {
-            if (rootDirectory.ContainsDirectory(".git"))
-                StartProcess("git", $"update-index --add --chmod=+x {scriptFile}", logInvocation: false, logOutput: false);
-
-            if (rootDirectory.ContainsDirectory(".svn"))
-                StartProcess("svn", $"propset svn:executable on {scriptFile}", logInvocation: false, logOutput: false);
-
-            if (IsUnix)
-                StartProcess("chmod", $"+x {scriptFile}", logInvocation: false, logOutput: false);
-        }
-    }
-
-    internal static void WriteConfigurationFile(AbsolutePath rootDirectory, AbsolutePath solutionFile)
-    {
-        var parametersFile = GetDefaultParametersFile(rootDirectory);
-        var dictionary = new Dictionary<string, string> { ["$schema"] = BuildSchemaFileName };
-        if (solutionFile != null)
-            dictionary["Solution"] = rootDirectory.GetUnixRelativePathTo(solutionFile).ToString();
-        parametersFile.WriteJson(dictionary, JsonExtensions.DefaultSerializerOptions);
-    }
-}
