+++ src/Fallout.Cli/Program.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Fallout.Cli.Commands;
+using Fallout.Cli.Commands.Navigation;
 using Fallout.Cli.Prompts;
 using Fallout.Common;
 using Fallout.Common.IO;
@@ -14,8 +15,6 @@ namespace Fallout.Cli;
 
 public partial class Program
 {
-    internal static string CurrentBuildScriptName => EnvironmentInfo.IsWin ? "build.ps1" : "build.sh";
-
     private static async Task<int> Main(string[] args)
     {
         Console.OutputEncoding = Encoding.UTF8;
@@ -25,7 +24,7 @@ public partial class Program
             var rootDirectory = TryGetRootDirectory();
 
             var buildScript = rootDirectory != null
-                ? rootDirectory.GetFiles(CurrentBuildScriptName, depth: 2)
+                ? rootDirectory.GetFiles(CliConventions.CurrentBuildScriptName, depth: 2)
                     .FirstOrDefault(x => Constants.TryGetRootDirectoryFrom(x.Parent) == rootDirectory)
                 : null;
 
@@ -44,6 +43,9 @@ public partial class Program
         var services = new ServiceCollection();
 
         services.AddSingleton<IConsolePrompts, SpectreConsolePrompts>();
+        services.AddSingleton<IConfigurationReader, ConfigurationReader>();
+        services.AddSingleton<IBuildScaffolder, BuildScaffolder>();
+        services.AddSingleton<IPackageManager, PackageManager>();
         services.AddSingleton<CommandDispatcher>();
         RegisterCommands(services);
 
@@ -52,30 +54,26 @@ public partial class Program
 
     private static void RegisterCommands(IServiceCollection services)
     {
-        // Real command types — issue #392 converts one legacy handler per PR.
         services.AddSingleton<IFalloutCommand, RunCommand>();
         services.AddSingleton<IFalloutCommand, TriggerCommand>();
+        services.AddSingleton<IFalloutCommand, CompleteCommand>();
+        services.AddSingleton<IFalloutCommand, GetConfigurationCommand>();
+        services.AddSingleton<IFalloutCommand, AddPackageCommand>();
+        services.AddSingleton<IFalloutCommand, UpdateCommand>();
+        services.AddSingleton<IFalloutCommand, SecretsCommand>();
+        services.AddSingleton<IFalloutCommand, CakeConvertCommand>();
+        services.AddSingleton<IFalloutCommand, CakeCleanCommand>();
+        services.AddSingleton<IFalloutCommand, GetNextDirectoryCommand>();
+        services.AddSingleton<IFalloutCommand, PopDirectoryCommand>();
+        services.AddSingleton<IFalloutCommand, PushWithCurrentRootDirectoryCommand>();
+        services.AddSingleton<IFalloutCommand, PushWithParentRootDirectoryCommand>();
+        services.AddSingleton<IFalloutCommand, PushWithChosenRootDirectoryCommand>();
 
-        // Legacy handlers still living on Program, adapted until they are extracted into command
-        // types. Each conversion deletes one line here plus its Program.X.cs partial.
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("setup", Setup));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("update", Update));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("add-package", AddPackage));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("cake-convert", CakeConvert));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("cake-clean", CakeClean));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("complete", Complete));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("get-configuration", GetConfiguration));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("secrets", Secrets));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("GetNextDirectory", (_, _, _) => GetNextDirectory()));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("PopDirectory", (_, _, _) => PopDirectory()));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("PushWithCurrentRootDirectory", (_, rootDirectory, _) => PushWithCurrentRootDirectory(rootDirectory)));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("PushWithParentRootDirectory", (_, rootDirectory, _) => PushWithParentRootDirectory(rootDirectory)));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("PushWithChosenRootDirectory", (_, _, _) => PushWithChosenRootDirectory()));
-    }
-
-    internal static void PrintInfo()
-    {
-        Host.Information($"Fallout Global Tool 🌐 {typeof(Program).Assembly.GetInformationalText()}");
+        // SetupCommand is also injected directly into CakeConvertCommand (cake conversion offers to
+        // scaffold a build first), so it is registered as a concrete singleton and surfaced as an
+        // IFalloutCommand through the same instance.
+        services.AddSingleton<SetupCommand>();
+        services.AddSingleton<IFalloutCommand>(sp => sp.GetRequiredService<SetupCommand>());
     }
 
     private static AbsolutePath TryGetRootDirectory()
@@ -90,19 +88,4 @@ public partial class Program
 
         return Constants.TryGetRootDirectoryFrom(Directory.GetCurrentDirectory());
     }
-
-    // ── Transitional Spectre prompt delegators ──────────────────────────────────────────────────
-    // The implementations now live in SpectreConsolePrompts; these forwards keep the not-yet-extracted
-    // Program.X.cs command handlers compiling. As each handler becomes a command type taking
-    // IConsolePrompts via the constructor, its use of these disappears; the last conversion deletes them.
-    private static readonly IConsolePrompts s_prompts = new SpectreConsolePrompts();
-
-    private static void ShowInput(string emoji, string title, string value) => s_prompts.ShowInput(emoji, title, value);
-    private static void ShowCompletion(string title) => s_prompts.ShowCompletion(title);
-    private static void ClearPreviousLine() => s_prompts.ClearPreviousLine();
-    private static bool PromptForConfirmation(string question) => s_prompts.PromptForConfirmation(question);
-    private static string PromptForInput(string question, string defaultValue = null) => s_prompts.PromptForInput(question, defaultValue);
-    private static string PromptForSecret(string title, int? minLength = null) => s_prompts.PromptForSecret(title, minLength);
-    private static T PromptForChoice<T>(string question, params (T Value, string Description)[] choices) => s_prompts.PromptForChoice(question, choices);
-    private static void ConfirmExecution(string title, Action action) => s_prompts.ConfirmExecution(title, action);
 }
